Move propagate_vocab_restrictions to opengever.base.restricted_vocab
