+++ opengever/base/behaviors/classification.py
@@ -1,7 +1,7 @@
 from five import grok
 from opengever.base import _
 from opengever.base.acquisition import set_default_with_acquisition
-from opengever.base.behaviors.utils import propagate_vocab_restrictions
+from opengever.base.restricted_vocab import propagate_vocab_restrictions
 from opengever.base.restricted_vocab import RestrictedVocabularyFactory
 from opengever.base.utils import language_cache_key
 from plone import api
