+++ opengever/base/behaviors/utils.py
@@ -1,54 +1,7 @@
 from plone.namedfile.utils import get_contenttype
 from urllib import quote
 from z3c.form.interfaces import HIDDEN_MODE
-from z3c.form.interfaces import IValue
-from z3c.form.value import ComputedValue
-from zope.component import getMultiAdapter
 import re
-
-
-def propagate_vocab_restrictions(container, event, restricted_fields, marker):
-
-    def dottedname(field):
-        return '.'.join((field.interface.__name__, field.__name__))
-
-    changed_fields = []
-    for desc in event.descriptions:
-        for name in desc.attributes:
-            changed_fields.append(name)
-
-    fields_to_check = []
-    for field in restricted_fields:
-        if dottedname(field) in changed_fields:
-            fields_to_check.append(field)
-
-    if not fields_to_check:
-        return
-
-    children = container.portal_catalog(
-        # XXX: Depth should not be limited (Issue #2027)
-        path={'depth': 2,
-              'query': '/'.join(container.getPhysicalPath())},
-        object_provides=(marker.__identifier__,)
-    )
-
-    for child in children:
-        obj = child.getObject()
-        for field in fields_to_check:
-            voc = field.bind(obj).source
-            value = field.get(field.interface(obj))
-            if value not in voc:
-                # obj, request, form, field, widget
-                default = getMultiAdapter((
-                    obj.aq_inner.aq_parent,
-                    obj.REQUEST,
-                    None,
-                    field,
-                    None,
-                ), IValue, name='default')
-                if isinstance(default, ComputedValue):
-                    default = default.get()
-                field.set(field.interface(obj), default)
 
 
 # Used as sortkey for sorting strings in numerical order
